open file with os/subprocess instead of webbrowser
webbrowser always tends to open the file verbatim in a browser tab. We
want the file to be opened with system default application like
"emacs (clicent)"

I tried many ways to configure firefox to open file:// link in
emacs(client) with out sucess. I applied this fix and it worked immediately.